# product-service/docs/swagger.yaml
definitions:
...
  request.RemoveImageRequest:
    properties:
      image_url:
        type: string
    type: object
...
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /admin/image:
    delete:
      consumes:
      - application/json
      description: Delete an image from storage by its URL
      parameters:
      - description: Remove Image Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RemoveImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request - URL required
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Remove image (admin/auth)
      tags:
      - images
  /admin/image-upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file to storage
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request - Invalid file or size
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Upload image (admin/auth)
      tags:
      - images
...
  /auth/image:
    delete:
      consumes:
      - application/json
      description: Delete an image from storage by its URL
      parameters:
      - description: Remove Image Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.RemoveImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request - URL required
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Remove image (admin/auth)
      tags:
      - images
  /auth/image-upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image file to storage
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/response.DefaultResponse'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request - Invalid file or size
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Upload image (admin/auth)
      tags:
      - images
...
swagger: "2.0"